Refactoring: ModelConductor module
